refactor: GsonExtensions
