+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/serialization/SerializationHelpers.kt
@@ -4,7 +4,7 @@ import com.google.gson.JsonDeserializationContext
 import com.google.gson.JsonElement
 import com.google.gson.JsonObject
 import com.google.gson.JsonSerializationContext
-import org.janelia.saalfeldlab.paintera.serialization.GsonExtensions.Companion.get
+import org.janelia.saalfeldlab.paintera.serialization.GsonExtensions.get
 
 object SerializationHelpers {
 	private const val TYPE_KEY = "type"
