+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/serialization/WindowProperties.kt
@@ -7,7 +7,7 @@ import javafx.beans.property.SimpleBooleanProperty
 import javafx.beans.property.SimpleIntegerProperty
 import javafx.beans.value.ObservableBooleanValue
 import org.janelia.saalfeldlab.fx.extensions.nonnull
-import org.janelia.saalfeldlab.paintera.serialization.GsonExtensions.Companion.get
+import org.janelia.saalfeldlab.paintera.serialization.GsonExtensions.get
 
 
 class WindowProperties {
